Move bootstrap logger tests next to factories
